ellipse.arima0 {ellipse}	R Documentation

Outline an approximate pairwise confidence region

Description

This function produces the ellipsoidal outline of an approximate pairwise confidence region for an ARIMA model fit.

Usage

## S3 method for class 'arima0'
ellipse(x, which = c(1, 2), level = 0.95, t = sqrt(qchisq(level, 2)), ...)

Arguments

x
The first argument should be an arima0 object, usually resulting from a call to arima0().

which
Which selects the pair of parameters to be plotted. The default is the first two.

level
The confidence level of the region. Default 95%.

t
The t statistic on the boundary of the ellipse.

...
Other ellipse.default parameters may also be used.

Details

The summary function is used to obtain the approximate covariance matrix of the fitted parameters.

Value

A matrix with columns x and y to outline the confidence region.

See Also

ellipse

Examples

data(USAccDeaths)
fit <- arima0(USAccDeaths, order = c(0, 1, 1), seasonal = list(order = c(0, 1, 1)))
# Plot the approximate 95% confidence region for the first two parameters
# of the model
plot(ellipse(fit), type = 'l')
points(fit$coef[1], fit$coef[2])

[Package ellipse version 0.5.0 Index]