ellipse.nls {ellipse}	R Documentation

Outline an approximate pairwise confidence region

Description

This function produces the ellipsoidal outline of an approximate pairwise confidence region for a nonlinear model fit.

Usage

## S3 method for class 'nls'
ellipse(x, which = c(1, 2), level = 0.95, 
    t = sqrt(2 * qf(level, 2, s$df[2])), ...)

Arguments

x
The first argument should be an nls object, usually resulting from a call to nls().

which
Which selects the pair of parameters to be plotted. The default is the first two.

level
The confidence level of the region. Default 95%.

t
The t statistic on the boundary of the ellipse.

...
Other ellipse.default parameters may also be used.

Details

The summary function is used to obtain the approximate covariance matrix of the fitted parameters.

Value

A matrix with columns x and y to outline the confidence region.

See Also

ellipse.default, ellipse.profile

Examples

# Plot an approximate 95% confidence region for the weight and displacement
# parameters in the Michaelis Menten model
data(Puromycin)
fit <- nls(rate ~ Vm*conc/(K + conc), data = Puromycin, subset = state=="treated", 
    start = list(K = 0.05, Vm = 200))
plot(ellipse(fit,which=c('Vm','K')), type = 'l')
params <- fit$m$getPars()
points(params['Vm'],params['K'])

[Package ellipse version 0.5.0 Index]